requirements.txt
pyserial==3.5
python-gsmmodem-new==0.13.0
requests==2.32.3